refactor(gateway): 게이트웨이 경로 재작성 필터 삭제
