+++ com.delivery-signal.eureka.client.gateway/src/main/java/com/delivery_signal/eureka/client/gateway/presentation/TokenExceptionHandler.java
@@ -5,11 +5,11 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import com.delivery_signal.eureka.client.gateway.domain.exception.EmptyClaimsException;
-import com.delivery_signal.eureka.client.gateway.domain.exception.ExpiredException;
-import com.delivery_signal.eureka.client.gateway.domain.exception.InvalidSignatureException;
-import com.delivery_signal.eureka.client.gateway.domain.exception.MalFormedException;
-import com.delivery_signal.eureka.client.gateway.domain.exception.UnsupportedException;
+import com.delivery_signal.eureka.client.gateway.global.exception.EmptyClaimsException;
+import com.delivery_signal.eureka.client.gateway.global.exception.ExpiredException;
+import com.delivery_signal.eureka.client.gateway.global.exception.InvalidSignatureException;
+import com.delivery_signal.eureka.client.gateway.global.exception.MalFormedException;
+import com.delivery_signal.eureka.client.gateway.global.exception.UnsupportedException;
 import com.delivery_signal.eureka.client.gateway.presentation.dto.ApiResponse;
 
 import lombok.extern.slf4j.Slf4j;
